refactor(ArticleController): session 체크 Controller로 이동 및 상세조회시 로그인된 사용자만 가능

+++ src/main/java/codesquad/codestagram/domain/article/ArticleController.java
@@ -1,6 +1,7 @@
 package codesquad.codestagram.domain.article;
 
 import codesquad.codestagram.common.constants.SessionConstants;
+import codesquad.codestagram.domain.auth.exception.UnauthorizedException;
 import codesquad.codestagram.domain.user.User;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
@@ -23,6 +24,10 @@ public class ArticleController {
                              @RequestParam String content,
                              HttpSession session) {
         User user = (User) session.getAttribute(SessionConstants.USER_SESSION_KEY);
+        if (user == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
         articleService.createArticle(title, content, user);
 
         return "redirect:/";
@@ -30,7 +35,14 @@ public class ArticleController {
 
     // 게시물 상세 조회
     @GetMapping("{id}")
-    public String viewArticle(@PathVariable Long id, Model model) {
+    public String viewArticle(@PathVariable Long id,
+                              Model model,
+                              HttpSession session) {
+        User user = (User) session.getAttribute(SessionConstants.USER_SESSION_KEY);
+        if (user == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
         Article article = articleService.findArticle(id);
         model.addAttribute("article", article);
 
@@ -42,6 +54,7 @@ public class ArticleController {
     public String writeArticle(HttpSession session) {
         User user = (User) session.getAttribute(SessionConstants.USER_SESSION_KEY);
         if (user == null) {
+            // 로그인 폼으로 이동
             return "redirect:/auth/login";
         }
 
@@ -77,6 +90,10 @@ public class ArticleController {
     public String deleteArticle(@PathVariable Long id,
                                 HttpSession session) {
         User user = (User) session.getAttribute(SessionConstants.USER_SESSION_KEY);
+        if (user == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
         articleService.deleteArticle(id, user);
 
         return "redirect:/";
